refactor: ロジックをカスタムフックにまとめる
