+++ frontend/src/features/sketches/SketchCreateForm/SketchCreateForm.tsx
@@ -1,0 +1,26 @@
+import { useRef } from "react";
+
+import { useSketchCreateForm } from "./useSketchCreateForm";
+
+import { SketchHandWriter } from "~/features/sketches/SketchHandWriter";
+import { Button } from "~/shared/components/Button";
+import { useCanvas } from "~/shared/hooks/useCanvas";
+
+export const SketchCreateForm = () => {
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+  const { clearCanvas } = useCanvas({ canvasRef });
+  const { createSketch } = useSketchCreateForm({ canvasRef });
+
+  return (
+    <div className="flex flex-col gap-8">
+      <div>
+        <SketchHandWriter canvasRef={canvasRef} />
+      </div>
+
+      <div className="flex flex-col gap-2">
+        <Button variant="secondary" onClick={clearCanvas}>リセット</Button>
+        <Button onClick={createSketch}>投稿する</Button>
+      </div>
+    </div>
+  );
+};
