+++ frontend/src/features/sketches/SketchCreateForm/useSketchCreateForm.ts
@@ -1,16 +1,14 @@
-import { useRef } from "react";
+import { RefObject } from "react";
 import { useNavigate } from "react-router-dom";
 
-import { SketchHandWriter } from "./SketchHandWriter";
-
-import { Button } from "~/shared/components/Button";
-import { useCanvas } from "~/shared/hooks/useCanvas";
 import { sketchApi } from "~/shared/services/API";
 
-export const SketchCreateForm = () => {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
+type Args = {
+  canvasRef: RefObject<HTMLCanvasElement>
+}
+
+export const useSketchCreateForm = ({ canvasRef }: Args) => {
   const navigate = useNavigate();
-  const { clearCanvas } = useCanvas({ canvasRef });
 
   const convertCanvasToFile = async (): Promise<File> => {
     const canvas = canvasRef.current;
@@ -38,17 +36,6 @@ export const SketchCreateForm = () => {
 
     navigate("/sketches");
   };
-  
-  return (
-    <div className="flex flex-col gap-8">
-      <div>
-        <SketchHandWriter canvasRef={canvasRef} />
-      </div>
 
-      <div className="flex flex-col gap-2">
-        <Button variant="secondary" onClick={clearCanvas}>リセット</Button>
-        <Button onClick={createSketch}>投稿する</Button>
-      </div>
-    </div>
-  );
+  return { createSketch };
 };
